requirements.txt
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
slack-sdk>=3.0.0
anthropic>=0.40.0
